comment section for event, frontend

+++ Gazebo/Gazebo/Models/Comment.cs
@@ -5,7 +5,7 @@ namespace Gazebo.Models
     public class Comment
     {
         public int CommentId { get; set; } 
-        public int PostGroupTypeId { get; set; }
+        public byte PostGroupTypeId { get; set; }
         public int PostGroupId { get; set; }
         public int CommentOwnerId { get; set; }
         public string CommentText { get; set; }
